pyproject.toml
[tool.poetry]
name = "jsonrpc"
version = "0.1.0"
description = ""
authors = ["TsMarch"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
pre-commit = "^4.0.1"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
django-types = "^0.19.1"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
profile = "black"
skip = ["JsonRpc/config/urls.py"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
